% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_missing.R
\name{plot_missing}
\alias{plot_missing}
\title{Plot Missing Values by Column}
\usage{
plot_missing(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
A ggplot2 object.
}
\description{
Plot Missing Values by Column
}
\examples{
plot_missing(airquality)
}
