% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_impute.R
\name{quick_impute}
\alias{quick_impute}
\title{Quickly Impute Missing Data}
\usage{
quick_impute(df)
}
\arguments{
\item{df}{A data frame containing missing values.}
}
\value{
A clean data frame with no missing values.
}
\description{
Automatically imputes missing values in a dataset. Numeric columns are imputed
using the median, and categorical columns are imputed using the mode.
}
\examples{
messy_df <- data.frame(
  age = c(25, NA, 30, 22),
  city = c("NY", "LA", NA, "NY")
)
clean_df <- quick_impute(messy_df)

}
